requirements.txt
anthropic>=0.30.0
Pillow>=10.0.0
mss>=9.0.0
python-dotenv>=1.0.0
pywin32>=306
psutil>=5.9.0
